Set lexical-binding explicitly
Emacs 30.0.50 has started to warn when this variable isn't set,
presumably so that the default can be changed from nil to t in
a few years.

It would be better to enable lexical-binding.

I also sneaked in a newline and removed a period to comply with
the summary line conventions.